+++ music/src/main/java/com/music/service/music/MusicUpdateService.java
@@ -1,0 +1,5 @@
+package com.music.service.music;
+
+public class MusicUpdateService {
+
+}
